renaming, moving files for a more sensible organizational structure
