+++ app/styles/components/TabView.styl
@@ -35,7 +35,7 @@ $tab-height = 30px;
   display: flex;
   align-items: center;
 
-  > .tab-container {
+  > ul.tabs {
     display: flex;
     height: 100%;
     align-items: flex-end;
